requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
qdrant-client==1.7.0
openai==1.12.0
tiktoken==0.5.2
python-dotenv==1.0.0

